fix(models): serialize observed cache publication
