+++ pkg/services/models/internal/services/assets/internal/service/generic_cache.go
@@ -162,479 +162,6 @@ func (s *service) inspectGenericCache(
 	return cached, missing, nil
 }
 
-func (s *service) inspectLegacyGenericCache(
-	ctx context.Context,
-	kind string,
-	source genericSource,
-	artifacts []genericArtifact,
-	roots []string,
-) (map[string]genericCachePath, bool, error) {
-	if len(roots) == 0 || !genericArtifactsHaveTrustedFacts(artifacts) {
-		return nil, false, nil
-	}
-	ownedRoot := roots[len(roots)-1]
-	records := s.discoverContentAddressedRecords(ownedRoot, kind, source)
-	var candidate *genericCacheRecord
-	for index := range records {
-		record := records[index]
-		if !isRepairableGenericCacheRecord(record, kind, source, artifacts) {
-			continue
-		}
-		if candidate != nil {
-			// Multiple stale records for the same source and artifact set do
-			// not identify a unique owner. Leave them untouched.
-			return nil, false, nil
-		}
-		candidate = &records[index]
-	}
-	if candidate == nil {
-		return nil, false, nil
-	}
-
-	cached := make(map[string]genericCachePath, len(artifacts))
-	for _, artifact := range artifacts {
-		path := filepath.Join(candidate.snapshotPath, filepath.FromSlash(artifact.requirement.Name))
-		found, ok, err := s.verifyGenericCachedArtifact(ctx, path, artifact.requirement)
-		if err != nil {
-			return cached, false, err
-		}
-		if !ok {
-			return nil, false, nil
-		}
-		found.legacySnapshotPath = candidate.snapshotPath
-		found.legacyMetadataPath = candidate.metadataPath
-		found.legacyRoot = candidate.root
-		found.legacyMetadata = candidate.metadata
-		cached[artifact.requirement.Name] = found
-	}
-	return cached, true, nil
-}
-
-func (s *service) verifyGenericCachedArtifact(
-	ctx context.Context,
-	path string,
-	requirement models.AssetRequirement,
-) (genericCachePath, bool, error) {
-	info, err := s.inspectPath(path)
-	if err != nil || info == nil || !info.Mode().IsRegular() {
-		return genericCachePath{}, false, nil
-	}
-	actual, err := s.fileSHA256(ctx, path)
-	if err != nil {
-		if contextErr := assetContextError(ctx); contextErr != nil {
-			return genericCachePath{}, false, contextErr
-		}
-		return genericCachePath{}, false, nil
-	}
-	if requirement.Bytes <= 0 || info.Size() != requirement.Bytes {
-		return genericCachePath{}, false, nil
-	}
-	expected := strings.ToLower(strings.TrimSpace(requirement.SHA256))
-	if expected == "" || actual != expected {
-		return genericCachePath{}, false, nil
-	}
-	return genericCachePath{
-		artifact: models.AssetArtifact{
-			Name: requirement.Name, Bytes: info.Size(), SHA256: actual,
-		},
-		path: path,
-	}, true, nil
-}
-
-func genericArtifactsHaveTrustedFacts(artifacts []genericArtifact) bool {
-	if len(artifacts) == 0 {
-		return false
-	}
-	for _, artifact := range artifacts {
-		if strings.TrimSpace(artifact.requirement.Name) == "" || artifact.requirement.Bytes <= 0 ||
-			strings.TrimSpace(artifact.requirement.SHA256) == "" {
-			return false
-		}
-	}
-	return true
-}
-
-func isRepairableGenericCacheRecord(
-	record genericCacheRecord,
-	kind string,
-	source genericSource,
-	artifacts []genericArtifact,
-) bool {
-	metadata := record.metadata
-	if !genericCacheMetadataNeedsRepair(metadata) || metadata.Kind != kind ||
-		metadata.Source != source.safe || metadata.SourceKey != genericSourceIdentity(source) ||
-		len(metadata.Artifacts) != len(artifacts) || len(metadata.Artifacts) == 0 {
-		return false
-	}
-	metadataArtifacts := genericArtifactsFromRequirements(metadata.Artifacts)
-	requested := genericArtifactRequirements(artifacts)
-	if genericCacheKey(kind, source, metadataArtifacts) != metadata.Identity ||
-		genericArtifactIdentityHash(kind, source, metadataArtifacts) != filepath.Base(record.snapshotPath) ||
-		!sameGenericArtifactNames(metadata.Artifacts, requested) ||
-		!legacyMetadataFactsAgree(metadata.Artifacts, requested) {
-		return false
-	}
-	return true
-}
-
-func genericCacheMetadataNeedsRepair(metadata genericCacheMetadata) bool {
-	if len(metadata.Artifacts) == 0 {
-		return false
-	}
-	for _, artifact := range metadata.Artifacts {
-		if artifact.Bytes < 0 {
-			return false
-		}
-		if artifact.Bytes == 0 || strings.TrimSpace(artifact.SHA256) == "" {
-			return true
-		}
-	}
-	return false
-}
-
-func genericArtifactsFromRequirements(requirements []models.AssetRequirement) []genericArtifact {
-	artifacts := make([]genericArtifact, 0, len(requirements))
-	for _, requirement := range requirements {
-		artifacts = append(artifacts, genericArtifact{requirement: requirement})
-	}
-	return artifacts
-}
-
-func genericArtifactRequirements(artifacts []genericArtifact) []models.AssetRequirement {
-	requirements := make([]models.AssetRequirement, 0, len(artifacts))
-	for _, artifact := range artifacts {
-		requirements = append(requirements, artifact.requirement)
-	}
-	return requirements
-}
-
-func sameGenericArtifactNames(first, second []models.AssetRequirement) bool {
-	if len(first) != len(second) {
-		return false
-	}
-	seen := make(map[string]struct{}, len(first))
-	for _, artifact := range first {
-		name := filepath.ToSlash(strings.TrimSpace(artifact.Name))
-		if !validGenericArtifactName(name) {
-			return false
-		}
-		if _, ok := seen[name]; ok {
-			return false
-		}
-		seen[name] = struct{}{}
-	}
-	for _, artifact := range second {
-		name := filepath.ToSlash(strings.TrimSpace(artifact.Name))
-		if !validGenericArtifactName(name) {
-			return false
-		}
-		if _, ok := seen[name]; !ok {
-			return false
-		}
-		delete(seen, name)
-	}
-	return len(seen) == 0
-}
-
-func legacyMetadataFactsAgree(
-	metadata []models.AssetRequirement,
-	requested []models.AssetRequirement,
-) bool {
-	requestedByName := make(map[string]models.AssetRequirement, len(requested))
-	for _, artifact := range requested {
-		requestedByName[filepath.ToSlash(strings.TrimSpace(artifact.Name))] = artifact
-	}
-	for _, artifact := range metadata {
-		name := filepath.ToSlash(strings.TrimSpace(artifact.Name))
-		requirement, ok := requestedByName[name]
-		if !ok {
-			return false
-		}
-		// A non-zero legacy fact is retained only when it agrees with the
-		// caller's complete requirement. Zero and blank values are the sole
-		// facts this compatibility path is allowed to fill.
-		if artifact.Bytes < 0 || (artifact.Bytes > 0 && artifact.Bytes != requirement.Bytes) {
-			return false
-		}
-		if digest := strings.TrimSpace(artifact.SHA256); digest != "" &&
-			!strings.EqualFold(digest, strings.TrimSpace(requirement.SHA256)) {
-			return false
-		}
-	}
-	return true
-}
-
-func validGenericArtifactName(name string) bool {
-	raw := strings.TrimSpace(name)
-	if raw == "" || strings.ContainsAny(raw, "\\\x00") {
-		return false
-	}
-	pathName := filepath.FromSlash(raw)
-	if filepath.IsAbs(pathName) || filepath.VolumeName(pathName) != "" {
-		return false
-	}
-	name = filepath.ToSlash(raw)
-	return name != "." && name != ".." && !strings.HasPrefix(name, "/") &&
-		filepath.ToSlash(filepath.Clean(filepath.FromSlash(name))) == name
-}
-
-func (s *service) repairLegacyGenericCache(
-	ctx context.Context,
-	kind string,
-	source genericSource,
-	artifacts []genericArtifact,
-	cached map[string]genericCachePath,
-	roots []string,
-) (map[string]genericCachePath, error) {
-	record, ok := legacyGenericCacheRecord(cached, artifacts)
-	if !ok || !genericArtifactsHaveTrustedFacts(artifacts) || len(roots) == 0 ||
-		filepath.Clean(record.root) != filepath.Clean(roots[len(roots)-1]) {
-		return cached, nil
-	}
-	if !isRepairableGenericCacheRecord(record, kind, source, artifacts) {
-		return cached, nil
-	}
-	observed, err := s.reverifyLegacyGenericArtifacts(ctx, record, artifacts)
-	if err != nil {
-		return cached, err
-	}
-	if err := assetContextError(ctx); err != nil {
-		return cached, err
-	}
-	identity := genericArtifactIdentityHash(kind, source, observed)
-	finalPath := filepath.Join(filepath.Dir(record.snapshotPath), identity)
-	if filepath.Clean(finalPath) == filepath.Clean(record.snapshotPath) {
-		return cached, fmt.Errorf(
-			"%w: legacy cache identity is not incomplete", models.ErrAssetIntegrityFailed,
-		)
-	}
-	if err := s.ensureLegacyRepairDestinationAbsent(finalPath); err != nil {
-		return cached, err
-	}
-	if err := assetContextError(ctx); err != nil {
-		return cached, err
-	}
-	backupPath, hadMetadata, err := s.replaceLegacyGenericMetadata(
-		record, kind, source, observed,
-	)
-	if err != nil {
-		return cached, err
-	}
-	if err := assetContextError(ctx); err != nil {
-		restoreErr := s.restoreLegacyMetadata(record.metadataPath, backupPath, hadMetadata)
-		return cached, joinLegacyRepairErrors(err, restoreErr)
-	}
-	if err := s.renamePath(record.snapshotPath, finalPath); err != nil {
-		restoreErr := s.restoreLegacyMetadata(record.metadataPath, backupPath, hadMetadata)
-		return cached, joinLegacyRepairErrors(
-			legacyRepairInterruption("move legacy cache snapshot", err), restoreErr,
-		)
-	}
-	if err := assetContextError(ctx); err != nil {
-		rollbackErr := s.rollbackLegacyRepair(finalPath, record.snapshotPath, backupPath, hadMetadata)
-		return cached, joinLegacyRepairErrors(err, rollbackErr)
-	}
-
-	repaired, err := s.reverifyRepairedGenericArtifacts(ctx, finalPath, observed)
-	if err != nil {
-		rollbackErr := s.rollbackLegacyRepair(finalPath, record.snapshotPath, backupPath, hadMetadata)
-		return cached, joinLegacyRepairErrors(err, rollbackErr)
-	}
-	if err := assetContextError(ctx); err != nil {
-		rollbackErr := s.rollbackLegacyRepair(finalPath, record.snapshotPath, backupPath, hadMetadata)
-		return cached, joinLegacyRepairErrors(err, rollbackErr)
-	}
-	if err := s.removeLegacyMetadataBackup(finalPath, backupPath, hadMetadata); err != nil {
-		return repaired, err
-	}
-	return repaired, nil
-}
-
-func (s *service) replaceLegacyGenericMetadata(
-	record genericCacheRecord,
-	kind string,
-	source genericSource,
-	observed []genericArtifact,
-) (string, bool, error) {
-	metadataStagePath := record.metadataPath + ".partial"
-	if err := s.removeTree(metadataStagePath); err != nil {
-		return "", false, legacyRepairInterruption("clear legacy metadata staging", err)
-	}
-	if err := s.writeGenericMetadata(
-		metadataStagePath, kind, genericCacheKey(kind, source, observed), source.safe,
-		genericSourceIdentity(source), observed,
-	); err != nil {
-		cleanupErr := s.removeTree(metadataStagePath)
-		return "", false, joinLegacyRepairErrors(
-			legacyRepairInterruption("stage legacy cache metadata", err), cleanupErr,
-		)
-	}
-	backupPath, hadMetadata, err := s.moveExistingGenericMetadata(record.metadataPath)
-	if err != nil {
-		cleanupErr := s.removeTree(metadataStagePath)
-		return "", false, joinLegacyRepairErrors(
-			legacyRepairInterruption("backup legacy cache metadata", err), cleanupErr,
-		)
-	}
-	if err := s.renamePath(metadataStagePath, record.metadataPath); err != nil {
-		restoreErr := s.restoreLegacyMetadata(record.metadataPath, backupPath, hadMetadata)
-		cleanupErr := s.removeTree(metadataStagePath)
-		return "", false, joinLegacyRepairErrors(
-			legacyRepairInterruption("replace legacy cache metadata", err),
-			joinLegacyRepairErrors(restoreErr, cleanupErr),
-		)
-	}
-	return backupPath, hadMetadata, nil
-}
-
-func (s *service) removeLegacyMetadataBackup(path, backupPath string, hadMetadata bool) error {
-	if !hadMetadata {
-		return nil
-	}
-	backup := filepath.Join(path, filepath.Base(backupPath))
-	if err := s.removePath(backup); err != nil {
-		return legacyRepairInterruption("clean legacy cache metadata backup", err)
-	}
-	return nil
-}
-
-func legacyGenericCacheRecord(
-	cached map[string]genericCachePath,
-	artifacts []genericArtifact,
-) (genericCacheRecord, bool) {
-	var record genericCacheRecord
-	for _, artifact := range artifacts {
-		found, ok := cached[artifact.requirement.Name]
-		if !ok || found.legacySnapshotPath == "" || found.legacyMetadataPath == "" {
-			return genericCacheRecord{}, false
-		}
-		if record.snapshotPath == "" {
-			record = genericCacheRecord{
-				root:         found.legacyRoot,
-				snapshotPath: found.legacySnapshotPath,
-				metadataPath: found.legacyMetadataPath,
-				metadata:     found.legacyMetadata,
-			}
-			continue
-		}
-		if record.snapshotPath != found.legacySnapshotPath || record.metadataPath != found.legacyMetadataPath {
-			return genericCacheRecord{}, false
-		}
-		if record.root != found.legacyRoot {
-			return genericCacheRecord{}, false
-		}
-	}
-	return record, record.snapshotPath != ""
-}
-
-func (s *service) reverifyLegacyGenericArtifacts(
-	ctx context.Context,
-	record genericCacheRecord,
-	artifacts []genericArtifact,
-) ([]genericArtifact, error) {
-	observed := make([]genericArtifact, 0, len(artifacts))
-	for _, artifact := range artifacts {
-		path := filepath.Join(record.snapshotPath, filepath.FromSlash(artifact.requirement.Name))
-		found, ok, err := s.verifyGenericCachedArtifact(ctx, path, artifact.requirement)
-		if err != nil {
-			return nil, err
-		}
-		if !ok {
-			return nil, fmt.Errorf(
-				"%w: legacy asset %q changed before repair", models.ErrAssetIntegrityFailed,
-				artifact.requirement.Name,
-			)
-		}
-		observed = append(observed, genericArtifact{
-			requirement: models.AssetRequirement{
-				Name:   artifact.requirement.Name,
-				Bytes:  found.artifact.Bytes,
-				SHA256: found.artifact.SHA256,
-			},
-			metadataResolved: true,
-		})
-	}
-	return observed, nil
-}
-
-func (s *service) ensureLegacyRepairDestinationAbsent(path string) error {
-	info, err := s.inspectPath(path)
-	if errors.Is(err, os.ErrNotExist) {
-		return nil
-	}
-	if err != nil {
-		return legacyRepairInterruption("inspect legacy cache destination", err)
-	}
-	if info == nil {
-		return legacyRepairInterruption("inspect legacy cache destination", errors.New("destination info unavailable"))
-	}
-	return fmt.Errorf("%w: legacy cache identity collision", models.ErrAssetIntegrityFailed)
-}
-
-func (s *service) reverifyRepairedGenericArtifacts(
-	ctx context.Context,
-	root string,
-	artifacts []genericArtifact,
-) (map[string]genericCachePath, error) {
-	repaired := make(map[string]genericCachePath, len(artifacts))
-	for _, artifact := range artifacts {
-		path := filepath.Join(root, filepath.FromSlash(artifact.requirement.Name))
-		found, ok, err := s.verifyGenericCachedArtifact(ctx, path, artifact.requirement)
-		if err != nil {
-			return repaired, err
-		}
-		if !ok {
-			return repaired, legacyRepairInterruption(
-				"validate repaired legacy cache artifact", errors.New("verified file changed"),
-			)
-		}
-		repaired[artifact.requirement.Name] = found
-	}
-	return repaired, nil
-}
-
-func (s *service) rollbackLegacyRepair(
-	currentPath string,
-	legacyPath string,
-	backupPath string,
-	hadMetadata bool,
-) error {
-	if err := s.renamePath(currentPath, legacyPath); err != nil {
-		return legacyRepairInterruption("restore legacy cache snapshot", err)
-	}
-	return s.restoreLegacyMetadata(filepath.Join(legacyPath, assetMetadataName), backupPath, hadMetadata)
-}
-
-func (s *service) restoreLegacyMetadata(path, backupPath string, hadMetadata bool) error {
-	if !hadMetadata {
-		return nil
-	}
-	if err := s.removePath(path); err != nil && !errors.Is(err, os.ErrNotExist) {
-		return legacyRepairInterruption("remove replacement legacy metadata", err)
-	}
-	if err := s.renamePath(backupPath, path); err != nil {
-		return legacyRepairInterruption("restore legacy cache metadata", err)
-	}
-	return nil
-}
-
-func legacyRepairInterruption(operation string, cause error) error {
-	return pullsupport.WrapPullStage(
-		models.PullStageCacheInstallation, "", operation, "",
-		interruptedAssetError(operation, cause),
-	)
-}
-
-func joinLegacyRepairErrors(primary, cleanup error) error {
-	if primary == nil {
-		return cleanup
-	}
-	if cleanup == nil {
-		return primary
-	}
-	return errors.Join(primary, legacyRepairInterruption("clean legacy cache repair", cleanup))
-}
-
 func (s *service) findGenericArtifact(
 	ctx context.Context,
 	kind string,
@@ -1171,13 +698,16 @@ func (s *service) lockGenericCache(
 	ctx context.Context,
 	kind string,
 	source genericSource,
-	artifacts []genericArtifact,
+	_ []genericArtifact,
 	roots []string,
 ) (io.Closer, error) {
 	if len(roots) == 0 {
 		return nil, nil
 	}
-	identity := genericArtifactIdentityHash(kind, source, artifacts)
+	// The observed artifact facts may differ from the request facts. Use a
+	// stable source-level owner before observation so incomplete and complete
+	// requests cannot publish the same observed snapshot concurrently.
+	identity := genericIdentityHash(kind, source, nil)
 	lockPath := filepath.Join(
 		roots[len(roots)-1], ".you-asset-locks", kind, identity+".lock",
 	)
