+++ pkg/services/models/internal/services/assets/internal/service/generic_cache_concurrency_test.go
@@ -468,6 +468,187 @@ func TestPrepareGenericAssetsRepairsLegacyCacheAcrossServices(t *testing.T) {
 	}
 }
 
+func TestPrepareGenericAssetsSerializesObservedIdentityPublishAcrossServices(t *testing.T) {
+	t.Parallel()
+
+	fixture := newObservedIdentityRaceFixture(t)
+	results := make(chan observedIdentityPreparation, 2)
+	startObservedIdentityPreparation(results, fixture.firstService, fixture.firstRequest)
+	<-fixture.transferStarted
+	startObservedIdentityPreparation(results, fixture.secondService, fixture.secondRequest)
+	<-fixture.lockState.secondAttempted
+	fixture.releaseTransfer()
+
+	prepared, reused := collectObservedIdentityPreparations(t, results, fixture.body, fixture.digest)
+	if prepared.result.Outcome != models.AssetPreparationPrepared ||
+		reused.result.Outcome != models.AssetPreparationAlreadyAvailable {
+		t.Fatalf("observed identity outcomes = %#v and %#v, want prepared then reused", prepared.result.Outcome, reused.result.Outcome)
+	}
+	assertObservedIdentityCoordination(t, fixture)
+	assertObservedIdentitySnapshot(t, fixture)
+}
+
+type observedIdentityPreparation struct {
+	result models.PrepareModelAssetsResult
+	err    error
+}
+
+type observedIdentityRaceFixture struct {
+	body            []byte
+	digest          string
+	cacheDirectory  string
+	reference       models.ModelReference
+	firstService    *service
+	secondService   *service
+	firstRequest    models.PrepareModelAssetsRequest
+	secondRequest   models.PrepareModelAssetsRequest
+	transferStarted <-chan struct{}
+	releaseTransfer func()
+	transfers       *atomic.Int32
+	lockState       *genericLockPathState
+}
+
+func newObservedIdentityRaceFixture(t *testing.T) *observedIdentityRaceFixture {
+	t.Helper()
+	body := []byte("observed identity concurrency payload")
+	digest := sha256Hex(body)
+	transferStarted := make(chan struct{})
+	releaseTransfer := make(chan struct{})
+	var transferOnce sync.Once
+	var releaseOnce sync.Once
+	transfers := &atomic.Int32{}
+	manifestClient := genericManifestWithoutDigestClient("weights.bin", func() []byte { return body })
+	client := httpDoerFunc(func(request *http.Request) (*http.Response, error) {
+		if request.Method == http.MethodGet && !strings.HasPrefix(request.URL.Path, "/models/") && transfers.Add(1) == 1 {
+			transferOnce.Do(func() { close(transferStarted) })
+			<-releaseTransfer
+		}
+		return manifestClient.Do(request)
+	})
+	release := func() { releaseOnce.Do(func() { close(releaseTransfer) }) }
+	t.Cleanup(release)
+
+	cacheDirectory := t.TempDir()
+	firstScopes := newScopes(t, "generic-observed-identity-first")
+	secondScopes := newScopes(t, "generic-observed-identity-second")
+	firstScope := openScope(t, firstScopes, cacheDirectory, models.RuntimeConfig{})
+	secondScope := openScope(t, secondScopes, cacheDirectory, models.RuntimeConfig{})
+	lockState := &genericLockPathState{secondAttempted: make(chan struct{})}
+	firstService := newGenericService(t, firstScopes, client, func(string) string { return "" })
+	secondService := newGenericService(t, secondScopes, client, func(string) string { return "" })
+	firstService.coordination = newGenericLockPathRecorder(t, lockState)
+	secondService.coordination = newGenericLockPathRecorder(t, lockState)
+	reference := models.ModelReference{NameOrURI: "hf://owner/repo/weights.bin@" + genericTestRevision}
+	return &observedIdentityRaceFixture{
+		body: body, digest: digest, cacheDirectory: cacheDirectory, reference: reference,
+		firstService: firstService, secondService: secondService,
+		firstRequest: models.PrepareModelAssetsRequest{
+			Scope: firstScope, Name: "observed-identity-first", Reference: reference,
+			Artifacts: []models.AssetRequirement{{Name: "weights.bin"}},
+		},
+		secondRequest: models.PrepareModelAssetsRequest{
+			Scope: secondScope, Name: "observed-identity-second", Reference: reference,
+			Artifacts: []models.AssetRequirement{{Name: "weights.bin", Bytes: int64(len(body)), SHA256: digest}},
+		},
+		transferStarted: transferStarted, releaseTransfer: release, transfers: transfers, lockState: lockState,
+	}
+}
+
+func newGenericLockPathRecorder(t *testing.T, state *genericLockPathState) platformlocking.Service {
+	t.Helper()
+	return &genericLockPathRecorder{delegate: mustPlatformLockingService(t), state: state}
+}
+
+func startObservedIdentityPreparation(
+	results chan<- observedIdentityPreparation,
+	assetService *service,
+	request models.PrepareModelAssetsRequest,
+) {
+	go func() {
+		result, err := assetService.PrepareModelAssets(context.Background(), request)
+		results <- observedIdentityPreparation{result: result, err: err}
+	}()
+}
+
+func collectObservedIdentityPreparations(
+	t *testing.T,
+	results <-chan observedIdentityPreparation,
+	body []byte,
+	digest string,
+) (observedIdentityPreparation, observedIdentityPreparation) {
+	t.Helper()
+	var prepared, reused observedIdentityPreparation
+	for index := 0; index < 2; index++ {
+		outcome := <-results
+		if outcome.err != nil {
+			t.Fatalf("observed identity preparation %d: %v", index, outcome.err)
+		}
+		assertObservedIdentityResult(t, outcome.result, body, digest, index)
+		switch outcome.result.Outcome {
+		case models.AssetPreparationPrepared:
+			prepared = outcome
+		case models.AssetPreparationAlreadyAvailable:
+			reused = outcome
+		default:
+			t.Fatalf("observed identity outcome %d = %v, want prepared or reused", index, outcome.result.Outcome)
+		}
+	}
+	return prepared, reused
+}
+
+func assertObservedIdentityResult(
+	t *testing.T,
+	result models.PrepareModelAssetsResult,
+	body []byte,
+	digest string,
+	index int,
+) {
+	t.Helper()
+	if result.Asset.Integrity != models.AssetIntegrityVerified || len(result.Asset.Artifacts) != 1 ||
+		result.Asset.Artifacts[0].Bytes != int64(len(body)) || result.Asset.Artifacts[0].SHA256 != digest {
+		t.Fatalf("observed identity result %d = %#v, want verified observed artifact", index, result)
+	}
+}
+
+func assertObservedIdentityCoordination(t *testing.T, fixture *observedIdentityRaceFixture) {
+	t.Helper()
+	if got := fixture.transfers.Load(); got != 1 {
+		t.Fatalf("observed identity transfers = %d, want one transfer", got)
+	}
+	if got := fixture.lockState.cacheCalls.Load(); got != 2 {
+		t.Fatalf("observed identity cache lock calls = %d, want two service owners", got)
+	}
+	fixture.lockState.mu.Lock()
+	lockPath := fixture.lockState.path
+	fixture.lockState.mu.Unlock()
+	if lockPath == "" {
+		t.Fatal("observed identity lock path is empty")
+	}
+}
+
+func assertObservedIdentitySnapshot(t *testing.T, fixture *observedIdentityRaceFixture) {
+	t.Helper()
+	source := genericSource{
+		kind: genericSourceHF, safe: fixture.reference.NameOrURI,
+		owner: "owner", repository: "repo", file: "weights.bin", revision: genericTestRevision,
+	}
+	artifacts := []genericArtifact{{requirement: models.AssetRequirement{
+		Name: "weights.bin", Bytes: int64(len(fixture.body)), SHA256: fixture.digest,
+	}}}
+	correctedPath := filepath.Join(
+		fixture.cacheDirectory, assetContentDirectory, assetKindModel,
+		genericArtifactIdentityHash(assetKindModel, source, artifacts),
+	)
+	assertGenericArtifact(t, filepath.Join(correctedPath, "weights.bin"), fixture.body, "observed identity committed artifact")
+	entries, err := os.ReadDir(filepath.Dir(correctedPath))
+	if err != nil {
+		t.Fatalf("read observed identity cache: %v", err)
+	}
+	if len(entries) != 1 || entries[0].Name() != filepath.Base(correctedPath) {
+		t.Fatalf("observed identity cache entries = %#v, want one committed snapshot", entries)
+	}
+}
+
 func requireGenericPreparationOutcome(
 	t *testing.T,
 	result models.PrepareModelAssetsResult,
@@ -541,6 +722,19 @@ type observingStagingCoordination struct {
 	once      sync.Once
 }
 
+type genericLockPathState struct {
+	mu              sync.Mutex
+	path            string
+	cacheCalls      atomic.Int32
+	secondAttempted chan struct{}
+	secondOnce      sync.Once
+}
+
+type genericLockPathRecorder struct {
+	delegate platformlocking.Service
+	state    *genericLockPathState
+}
+
 func mustPlatformLockingService(t *testing.T) platformlocking.Service {
 	t.Helper()
 	service, err := platformlocking.New(platformlocking.LocalFileSystem{})
@@ -555,10 +749,204 @@ func (coordination *observingStagingCoordination) Lock(ctx context.Context, path
 	return coordination.delegate.Lock(ctx, path)
 }
 
+func (coordination *genericLockPathRecorder) Lock(ctx context.Context, path string) (io.Closer, error) {
+	isRuntimeLock := strings.Contains(filepath.ToSlash(path), "/.you-asset-locks/runtime/")
+	if !isRuntimeLock && coordination.state.cacheCalls.Add(1) == 2 {
+		coordination.state.secondOnce.Do(func() { close(coordination.state.secondAttempted) })
+	}
+	coordination.state.mu.Lock()
+	if !isRuntimeLock {
+		if coordination.state.path == "" {
+			coordination.state.path = path
+		} else if coordination.state.path != path {
+			coordination.state.mu.Unlock()
+			return nil, errors.New("generic cache ownership key changed with artifact facts")
+		}
+	}
+	coordination.state.mu.Unlock()
+	return coordination.delegate.Lock(ctx, path)
+}
+
 type rejectingStagingCoordination struct {
 	err error
 }
 
 func (coordination rejectingStagingCoordination) Lock(context.Context, string) (io.Closer, error) {
 	return nil, coordination.err
+}
+
+func TestGenericPrivatePlanningAndCachePreservation(t *testing.T) {
+	t.Parallel()
+
+	scopes := newScopes(t, "generic-private-planning")
+	service := newGenericService(t, scopes, httpDoerFunc(func(*http.Request) (*http.Response, error) {
+		return nil, errors.New("private planning test must not use HTTP")
+	}), func(string) string { return "" })
+
+	assertGenericLocalRequirements(t, service)
+	assertGenericOverlayPlanning(t)
+	assertGenericCachePreservation(t, service)
+}
+
+func assertGenericLocalRequirements(t *testing.T, service *service) {
+	t.Helper()
+
+	localRoot := t.TempDir()
+	if err := os.MkdirAll(filepath.Join(localRoot, "nested"), 0o755); err != nil {
+		t.Fatalf("create local fixture: %v", err)
+	}
+	if err := os.WriteFile(filepath.Join(localRoot, "root.bin"), []byte("root"), 0o644); err != nil {
+		t.Fatalf("write root fixture: %v", err)
+	}
+	if err := os.WriteFile(filepath.Join(localRoot, "nested", "child.bin"), []byte("child"), 0o644); err != nil {
+		t.Fatalf("write nested fixture: %v", err)
+	}
+	requirements, err := service.localRequirements(localRoot)
+	if err != nil || len(requirements) != 2 || requirements[0].Name != "nested/child.bin" || requirements[1].Name != "root.bin" {
+		t.Fatalf("localRequirements = %#v, %v, want sorted recursive files", requirements, err)
+	}
+	if _, err := service.localRequirements(filepath.Join(localRoot, "missing")); !errors.Is(err, models.ErrAssetSourceMissing) {
+		t.Fatalf("missing localRequirements error = %v, want ErrAssetSourceMissing", err)
+	}
+}
+
+func assertGenericOverlayPlanning(t *testing.T) {
+	t.Helper()
+
+	sourceText := "hf://owner/repo@" + genericTestRevision
+	overlays := map[string]models.ModelOverlay{
+		"Alias":  {Source: &sourceText},
+		"alias ": {Source: func() *string { value := "hf://owner/other@" + genericTestRevision; return &value }()},
+	}
+	name, overlay, ok := genericOverlay(overlays, "alias")
+	if !ok || name != "Alias" || overlay.Source == nil {
+		t.Fatalf("genericOverlay = %q, %#v, %v", name, overlay, ok)
+	}
+	*overlay.Source = "mutated"
+	if *overlays["Alias"].Source != sourceText {
+		t.Fatal("genericOverlay returned a mutable source pointer")
+	}
+	safe := genericHFSafeReference(genericSource{owner: "owner", repository: "repo", file: "weights.bin", revision: genericTestRevision})
+	if safe != "hf://owner/repo/weights.bin@"+genericTestRevision {
+		t.Fatalf("genericHFSafeReference = %q", safe)
+	}
+	var revisionFailure *models.InvocationFailure
+	if !errors.As(genericRevisionFailure(), &revisionFailure) || revisionFailure.Class != models.InvocationFailureClassRevisionResolution {
+		t.Fatalf("genericRevisionFailure = %#v", genericRevisionFailure())
+	}
+}
+
+func assertGenericCachePreservation(t *testing.T, service *service) {
+	t.Helper()
+
+	cachedPath := filepath.Join(t.TempDir(), "cached.bin")
+	if err := os.WriteFile(cachedPath, []byte("cached"), 0o644); err != nil {
+		t.Fatalf("write cached fixture: %v", err)
+	}
+	stagePath := t.TempDir()
+	artifact := genericCachePath{
+		artifact: models.AssetArtifact{Name: "nested/cached.bin", Bytes: int64(len("cached"))},
+		path:     cachedPath,
+	}
+	if err := service.preserveGenericArtifact(context.Background(), artifact, artifact.artifact.Name, stagePath); err != nil {
+		t.Fatalf("preserveGenericArtifact: %v", err)
+	}
+	preserved, err := os.ReadFile(filepath.Join(stagePath, "nested", "cached.bin"))
+	if err != nil || string(preserved) != "cached" {
+		t.Fatalf("preserved cached artifact = %q, %v", preserved, err)
+	}
+}
+
+func TestGenericSnapshotDiscoveryAndSourcePlanning(t *testing.T) {
+	t.Parallel()
+
+	scopes := newScopes(t, "generic-snapshot-planning")
+	service := newGenericService(t, scopes, httpDoerFunc(func(*http.Request) (*http.Response, error) {
+		return nil, errors.New("source planning test must not use HTTP")
+	}), func(string) string { return "" })
+	assertGenericSnapshotDiscovery(t, service)
+	assertGenericCacheRoots(t, service)
+	assertGenericSourceResolution(t, service)
+	assertGenericPreparationSelection(t)
+}
+
+func assertGenericSnapshotDiscovery(t *testing.T, service *service) {
+	t.Helper()
+	snapshotRoot := t.TempDir()
+	if err := os.MkdirAll(filepath.Join(snapshotRoot, "nested"), 0o755); err != nil {
+		t.Fatalf("create snapshot fixture: %v", err)
+	}
+	for name, body := range map[string]string{
+		"root.bin":            "root",
+		"nested/child.bin":    "child",
+		".hidden/ignored.bin": "ignored",
+	} {
+		path := filepath.Join(snapshotRoot, filepath.FromSlash(name))
+		if err := os.MkdirAll(filepath.Dir(path), 0o755); err != nil {
+			t.Fatalf("create snapshot parent: %v", err)
+		}
+		if err := os.WriteFile(path, []byte(body), 0o644); err != nil {
+			t.Fatalf("write snapshot file: %v", err)
+		}
+	}
+	requirements := service.discoverSnapshotRequirements(snapshotRoot)
+	if len(requirements) != 2 || requirements[0].Name != "nested/child.bin" || requirements[1].Name != "root.bin" {
+		t.Fatalf("discoverSnapshotRequirements = %#v, want visible sorted files", requirements)
+	}
+	if got := service.discoverSnapshotRequirements(filepath.Join(snapshotRoot, "missing")); got != nil {
+		t.Fatalf("missing snapshot requirements = %#v, want nil", got)
+	}
+}
+
+func assertGenericCacheRoots(t *testing.T, service *service) {
+	t.Helper()
+	scopeConfig := models.RuntimeScopeConfig{CacheDirectory: t.TempDir()}
+	modelRoots, err := service.genericCacheRoots(scopeConfig, models.AssetArtifactKindModel)
+	if err != nil || len(modelRoots) == 0 || modelRoots[len(modelRoots)-1] != scopeConfig.CacheDirectory {
+		t.Fatalf("model genericCacheRoots = %#v, %v", modelRoots, err)
+	}
+	backendRoots, err := service.genericCacheRoots(scopeConfig, models.AssetArtifactKindBackend)
+	if err != nil || len(backendRoots) != 1 || !strings.HasSuffix(filepath.ToSlash(backendRoots[0]), "/backend-artifacts") {
+		t.Fatalf("backend genericCacheRoots = %#v, %v", backendRoots, err)
+	}
+}
+
+func assertGenericSourceResolution(t *testing.T, service *service) {
+	t.Helper()
+	scopeConfig := models.RuntimeScopeConfig{CacheDirectory: t.TempDir()}
+	service.resolveRevision = func(context.Context, string) (string, error) { return genericTestRevision, nil }
+	resolved, err := service.resolveGenericSource(context.Background(), scopeConfig, "hf://owner/repo")
+	if err != nil || resolved.revision != genericTestRevision || resolved.safe != "hf://owner/repo@"+genericTestRevision {
+		t.Fatalf("resolved generic source = %#v, %v", resolved, err)
+	}
+	service.resolveRevision = func(context.Context, string) (string, error) { return "not-a-commit", nil }
+	if _, err := service.resolveGenericSource(context.Background(), scopeConfig, "hf://owner/repo"); !errors.Is(err, models.ErrModelRevisionUnresolved) {
+		t.Fatalf("unresolved revision error = %v, want ErrModelRevisionUnresolved", err)
+	}
+	if _, err := parseGenericReleaseSource("https://example.com/releases/download/v1/backend.tar"); !errors.Is(err, models.ErrModelReferenceInvalid) {
+		t.Fatalf("invalid release source error = %v, want ErrModelReferenceInvalid", err)
+	}
+	release, err := parseGenericReleaseSource("https://github.com/owner/repo/releases/download/v1/backend.tar")
+	if err != nil || release.kind != genericSourceRelease || release.artifactURL == "" {
+		t.Fatalf("valid release source = %#v, %v", release, err)
+	}
+}
+
+func assertGenericPreparationSelection(t *testing.T) {
+	t.Helper()
+	for _, request := range []models.PrepareModelAssetsRequest{
+		{Reference: models.ModelReference{NameOrURI: "hf://owner/repo@" + genericTestRevision}},
+		{Offline: true},
+		{Artifacts: []models.AssetRequirement{}},
+		{Backend: "backend-v1"},
+		{Name: "llm"},
+		{Name: "./local/model"},
+	} {
+		if !shouldPrepareGenericAssets(request) {
+			t.Fatalf("shouldPrepareGenericAssets(%#v) = false, want true", request)
+		}
+	}
+	if shouldPrepareGenericAssets(models.PrepareModelAssetsRequest{Name: "unknown-symbol"}) {
+		t.Fatal("unknown symbolic name should not select generic preparation")
+	}
 }
